requirements.txt
asgiref==3.8.1
Django==5.0.4
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-js-asset==2.2.0
sqlparse==0.4.4
